fix: constrain detail-screen header titles so long names stop overflowing
The scroll-linked header title on food/exam/lecture/lecturer/sports/CL
event detail screens is a fully custom headerTitle render prop, so it
gets no native truncation against the header's leading/trailing
buttons. A long name (e.g. a long meal or lecture title) painted past
the close button instead of being clipped.

Extract the identical six-screen JSX into one shared
DetailStackHeader (src/components/Universal), add an explicit maxWidth
that reserves space for the header's side buttons (getPlatformHeaderButtons
never renders more than one per side), and make the truncation intent
explicit with ellipsizeMode="tail" alongside HeaderTitle's existing
numberOfLines default.
