+++ src/components/Food/meal-detail-header.tsx
@@ -1,15 +1,10 @@
-import { HeaderTitle } from '@react-navigation/elements'
-import { Stack } from 'expo-router'
 import type { ViewStyle } from 'react-native'
-import { Platform, View } from 'react-native'
-import Animated, {
+import {
 	interpolate,
 	useAnimatedScrollHandler,
 	useAnimatedStyle,
 	useSharedValue
 } from 'react-native-reanimated'
-import { useCSSVariable } from 'uniwind'
-import { toColor } from '@/utils/uniwind-utils'
 
 export function useMealDetailScroll(): {
 	scrollHandler: ReturnType<typeof useAnimatedScrollHandler>
@@ -40,38 +35,4 @@ export function useMealDetailScroll(): {
 	})
 
 	return { scrollHandler, headerStyle }
-}
-
-interface MealDetailStackHeaderProps {
-	title: string
-	headerStyle: ReturnType<typeof useAnimatedStyle<ViewStyle>>
-}
-
-export function MealDetailStackHeader({
-	title,
-	headerStyle
-}: MealDetailStackHeaderProps): React.JSX.Element {
-	const textColor = toColor(useCSSVariable('--color-text'))
-
-	return (
-		<Stack.Screen
-			options={{
-				headerTitle: (props) => (
-					<View
-						className="overflow-hidden"
-						style={{
-							marginBottom: Platform.OS === 'ios' ? -10 : 0,
-							paddingRight: Platform.OS === 'ios' ? 0 : 50
-						}}
-					>
-						<Animated.View style={headerStyle}>
-							<HeaderTitle {...props} tintColor={String(textColor)}>
-								{title}
-							</HeaderTitle>
-						</Animated.View>
-					</View>
-				)
-			}}
-		/>
-	)
 }
